fix icc warning.
git-svn-id: http://root.cern.ch/svn/root/trunk@30774 27541ba8-7e3a-0410-8455-c3a389f83636

+++ graf3d/x3d/src/x3d.c
@@ -501,7 +501,7 @@ int x = 5, y = 5;
    XSetFillStyle(g->dpy, g->helpGc, FillOpaqueStippled);
 
    DrawLogo(g, (XTextWidth(g->font, LONGESTSTRING, strlen(LONGESTSTRING)) /
-   strlen(LONGESTSTRING)) / 2, FONTHEIGHT(g->font) / 3);
+   (int)strlen(LONGESTSTRING)) / 2, FONTHEIGHT(g->font) / 3);
 
    XSetFillStyle(g->dpy, g->helpGc, FillSolid);
